+++ src/search.js
@@ -6,11 +6,7 @@ exports.prepareSearchUrls = async ({ lat, lng, zoom, country, state, city, posta
     // Base part of the URLs to make up the startRequests
     const startUrlSearches = [];
 
-    // save geolocation to keyval
-    let geo = await Apify.getValue('GEO');
-    Apify.events.on('migrating', async () => {
-        await Apify.setValue('GEO', geo);
-    });
+    let geo;
 
     // preference for startUrlSearches is lat & lng > & state & city
     if (lat || lng) {
@@ -19,8 +15,13 @@ exports.prepareSearchUrls = async ({ lat, lng, zoom, country, state, city, posta
         }
         startUrlSearches.push(`https://www.google.com/maps/@${lat},${lng},${zoom}z/search`);
     } else if (country || state || city || postalCode) {
-        // Takes from KV or crate new one
+        geo = await Apify.getValue('GEO');
+        // Takes from KV or create a new one
         geo = geo || await getGeolocation({ country, state, city, postalCode });
+
+        Apify.events.on('migrating', async () => {
+            await Apify.setValue('GEO', geo);
+        });
 
         const points = await findPointsInPolygon(geo, zoom);
         for (const point of points) {